if linting issue

+++ src/fcmath/core/plots.py
@@ -30,7 +30,12 @@ class PlotRange:
         if self.minimum >= self.maximum:
             msg = "minimum must be smaller than maximum"
             raise ValueError(msg)
-        return np.linspace(self.minimum, self.maximum, self.samples)
+        return np.linspace(
+            self.minimum,
+            self.maximum,
+            self.samples,
+            dtype=np.float64,
+        )
 
 
 @dataclass(frozen=True)
